requirements.txt
aiohttp==3.10.11
aiohttp_socks==0.10.1
async-timeout==4.0.2
attrs==23.1.0
python-dotenv==1.0.0
websockets==13.1